backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
PyJWT==2.9.0
pytest==8.3.2
httpx==0.27.0
celery==5.4.0
redis==5.0.8
psycopg[binary]==3.2.1
